implemented the logic for handling all requests

+++ app/Http/Controllers/TicketController.php
@@ -81,7 +81,7 @@ class TicketController extends Controller
     public function update_2(TicketRequest $request) {
         $ticket = Ticket::find($request->id);
 
-        if ($request->equipment_id == 0) {
+        if ($request->equipment_id == '0') {
             $request['equipment_id'] = null;
             $request['status_id'] = Ticket::WAITING_FOR_EQUIPMENT;
         } 
@@ -106,7 +106,7 @@ class TicketController extends Controller
         // dd($request);
         $ticket = Ticket::find($request->id);
 
-        if ($request->HR_approval == Ticket::REJECTED) {
+        if ($ticket->ticket_type == Ticket::NEW_EQUIPMENT && $ticket->equipment_id != null && $request->HR_approval == Ticket::REJECTED) {
             $ticket->equipment->update(['available_quantity' => $ticket->equipment->available_quantity + 1]);
             $ticket->reservation->delete();
         }
@@ -118,12 +118,55 @@ class TicketController extends Controller
     public function update_4(TicketRequest $request) {
         $ticket = Ticket::find($request->id);
 
+        // dd($ticket);
         // dd($request);
 
         // if this a new equipment request and it has an equipment id, it means we only need 
         // to create a new document and add the reserved equipment
-        if ($ticket->ticket_type == Ticket::NEW_EQUIPMENT) {
-            if ($ticket->equipment_id != null && $ticket->HR_approval == Ticket::APPROVED) {
+        if ($ticket->isNewEquipmentRequest()) {
+            if ($ticket->equipment_id != null && $ticket->HR_approval == Ticket::APPROVED || $ticket->officer_approval == Ticket::REJECTED && $ticket->HR_approval == Ticket::APPROVED) {
+                // dd('here1');
+                $new_doc = Document::create([
+                    'user_id' => $ticket->user_id,
+                    'admin_id' => $ticket->officer_id,
+                    'date' => Carbon::now()->timestamp
+                ]);
+        
+                if ($ticket->equipment_id != null) {
+                    $equipment_id = $ticket->equipment_id;
+                } else {
+                    $equipment_id = $request['equipment_id'];
+                }
+
+                if ($ticket->serial_number_id != null) {
+                    $serial_number_id = $ticket->serial_number_id;
+                } else {
+                    $serial_number_id = $request['serial_number_id'];
+                }
+
+                DocumentItem::create([
+                    'document_id' => $new_doc->id,
+                    'equipment_id' => $equipment_id,
+                    'serial_number_id' => $serial_number_id            
+                ]);
+
+                if ($ticket->reservation != null) {
+                    // delete the reservation and stop counting it
+                    $ticket->reservation->delete(); 
+                    $ticket->equipment->update(['available_quantity' => $ticket->equipment->available_quantity + 1]);
+                }
+
+                $request['document_id'] = $new_doc->id; 
+
+                $ticket->update($request->only(['status_id', 'date_finished', 'serial_number_id', 'equipment_id', 'document_id']));
+
+                $ticket->equipment->update(['available_quantity' => $ticket->equipment->available_quantity - 1]);
+                
+                // if we had waited for equipment to arrive and have HR's approval
+                // list out the available equipment so that the officer can assign it to the user and close this request  
+            } else if ($ticket->status_id == Ticket::WAITING_FOR_EQUIPMENT && $ticket->HR_approval == Ticket::APPROVED) {
+                
+                // dd('here2');
                 $new_doc = Document::create([
                     'user_id' => $ticket->user_id,
                     'admin_id' => $ticket->officer_id,
@@ -132,20 +175,25 @@ class TicketController extends Controller
         
                 DocumentItem::create([
                     'document_id' => $new_doc->id,
-                    'equipment_id' => $ticket->equipment_id,
-                    'serial_number_id' => $ticket->serial_number_id            
+                    'equipment_id' => $request->equipment_id,
+                    'serial_number_id' => $request->serial_number_id            
                 ]);
-    
-                // delete the reservation
-                $ticket->reservation->delete();  
-                // if we had waited for equipment to arrive and have HR's approval
-                // list out the available equipment so that the officer can assign it to the user and close this request  
-            } else if ($ticket->status_id == Ticket::WAITING_FOR_EQUIPMENT && $ticket->HR_approval == Ticket::APPROVED) {
 
-            } 
+                $request['document_id'] = $new_doc->id;
+
+                $ticket->update($request->only(['status_id', 'date_finished', 'serial_number_id', 'equipment_id', 'document_id']));
+                $ticket->equipment->update(['available_quantity' => $ticket->equipment->available_quantity - 1]);
+
+            } else {
+                // dd('here3');
+                $ticket->update($request->only(['status_id', 'date_finished', 'serial_number_id', 'equipment_id', 'document_id']));
+            }
+
+        } else {
+            // dd('here4');
+            $ticket->update($request->only(['status_id', 'date_finished', 'serial_number_id', 'equipment_id', 'document_id']));
         }
 
-        $ticket->update($request->only(['status_id', 'date_finished', 'serial_number_id', 'equipment_id']));
         return redirect()->back();
     }
 }
